Refactor bot into modules
