+++ .env.example
@@ -1,0 +1,3 @@
+API_TOKEN=your_api_token
+CREDENTIALS_FILE=path/to/credentials.json
+ADMIN_IDS=123456789
